move QWebEngine classes to ui library
